dangdai-api/pyproject.toml
[project]
name = "dangdai-api"
version = "0.1.0"
description = "Python backend for Dangdai Chinese learning app with LangGraph quiz generation"
authors = [
    { name = "Maxime" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "langgraph>=1.0.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "python-dotenv>=1.0.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "supabase>=2.0.0",
    "pydantic-settings>=2.0.0",
    "PyJWT>=2.8.0",
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1", "pytest>=8.0.0"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "httpx>=0.28.1",
    "langgraph-cli[inmem]>=0.4.7",
    "mypy>=1.13.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.2",
]
